dataset-finetuning/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
pydantic>=1.10
sentence-transformers>=2.2.2
transformers>=4.40.0
torch>=2.0.0
faiss-cpu>=1.7.4.post2
pandas>=1.5
numpy>=1.24
python-multipart>=0.0.6
requests>=2.31
fhir.resources>=6.0.0
sqlmodel>=0.0.8
python-jose>=3.3.0
scikit-learn>=1.2.2
openpyxl>=3.1.0
